[Feat] 인사캠 버전 추가
