+++ src/App.tsx
@@ -1,7 +1,8 @@
-import React from "react";
 import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import MainPage from "@pages/Main";
+import SelectPage from "@pages/SelectPage";
+import JaGwaPage from "@pages/JaGwaPage";
+import InSaPage from "@pages/InSaPage";
 import { Analytics } from "@vercel/analytics/react";
 
 function App() {
@@ -10,7 +11,9 @@ function App() {
       <div className="App">
         <BrowserRouter>
           <Routes>
-            <Route path="/" element={<MainPage />}></Route>
+            <Route path="/" element={<SelectPage />}></Route>
+            <Route path="/jagwa" element={<JaGwaPage />}></Route>
+            <Route path="/insa" element={<InSaPage />}></Route>
           </Routes>
         </BrowserRouter>
       </div>
