+++ src/assets/data/data.json
@@ -96,5 +96,105 @@
     ["큰맘할매순대국", "37373939", 8000, 237000],
     ["토마스매운치즈등갈비", "35782198", 12000, 12000]
   ],
-  "insa": []
+  "insa": [
+    ["벅벅", "1652144509", 6700, 13000],
+    ["홍순두부", "1247371776", 6000, 7000],
+    ["동네식당", "1420402190", 4000, 6000],
+    ["김밥친구", "18678747", 2000, 3000],
+    ["더닭 성대점", "33067936", 8000, 11000],
+    ["뽀글뚝배기냠냠비빔밥", "1190976767", 6500, 7000],
+    ["벤프", "37432263", 6400, 8900],
+    ["밀가", "1039535707", 5500, 10000],
+    ["또와분식", "1108010747", 7000, 8500],
+    ["동네식당", "1420402190", 4000, 6000],
+    ["목우촌숯불갈비", "18678736", 7000, 29000],
+    ["꿀맛백반", "34457482", 8000, 30000],
+    ["진공푸", "1085865770", 9500, 12500],
+    ["은행골식당", "1157153754", 8500, 13800],
+    ["일송칼국수", "13534204", 8000, 9500],
+    ["한국관", "18677688", 12000, 35000],
+    ["노말키친", "1390630768", 14000, 17000],
+    ["명륜진사갈비", "1233726670", 16500, 16500],
+    ["신장군마라탕", "1550853249", 10000, 28000],
+    ["한솥도시락", "13376699", 4300, 7800],
+    ["페르시안궁전", "11710271", 14000, 33000],
+    ["타겟", "1873388687", 6000, 37500],
+    ["계략가", "1344431180", 23000, 23000],
+    ["피자헤븐", "1441859399", 18900, 33900],
+    ["맘스터치 피자앤치킨", "1721452122", 14600, 23000],
+    ["마마그릭", "1866966139", 5400, 12400],
+    ["성균원", "17975753", 6000, 21000],
+    ["간짜장전문점세아", "1939730046", 7000, 15000],
+    ["성대국수", "38275701", 8000, 8000],
+    ["이삭토스트", "38227457", 3500, 4600],
+    ["싸움의고수", "1142924239", 7500, 11000],
+    ["성균관감자탕", "31689555", 7000, 23000],
+    ["명쭈3", "247671846", 27000, 45000],
+    ["환식당", "1221474374", 6900, 15900],
+    ["천향록", "38521890", 12000, 35000],
+    ["나누미떡볶이", "11681136", 5000, 5000],
+    ["다발김치찌개", "1998666930", 7000, 7000],
+    ["또오리", "12048016", 17000, 84000],
+    ["온프", "1904231959", 9900, 14900],
+    ["발리밥", "1972198442", 11000, 14000],
+    ["도스타코스", "30853050", 8000, 10500],
+    ["가꾸스시", "1118442855", 11000, 36000],
+    ["포보", "13561663", 6500, 7500],
+    ["행복은간장밥", "1381351274", 9900, 10900],
+    ["미분당", "408557767", 9500, 12000],
+    ["뎁짜이", "34431003", 8000, 13000],
+    ["깔리", "12765748", 9000, 17000],
+    ["이유식당", "1460401099", 9000, 10500],
+    ["창화당", "1574394813", 7000, 11000],
+    ["히메시야", "1862295844", 9500, 15500],
+    ["멘야산다이메", "20322561", 9000, 13000],
+    ["은행골", "37746744", 12000, 34000],
+    ["호호식당", "85806567", 15000, 21000],
+    ["오이지", "1791259060", 9800, 26000],
+    ["나래함박", "1287736396", 15000, 21000],
+    ["은식당", "1193580592", 11000, 13000],
+    ["갸우뚱", "1854089051", 11800, 35000],
+    ["서양집", "717158780", 13900, 32900],
+    ["부탄츄", "83104110", 9000, 10000],
+    ["육미안", "1718133788", 9000, 69000],
+    ["마녀주방", "1663168040", 7500, 13900],
+    ["유자유김치떡볶이", "1106282746", 13500, 21000],
+    ["새로오픈장인닭갈비", "1921219764", 12000, 12000],
+    ["군자대한곱창", "1799831836", 17900, 22900],
+    ["혜화도담", "1638640485", 10000, 40000],
+    ["한사발", "1754181863", 8000, 33000],
+    ["돈돈정", "37236150", 12000, 18000],
+    ["야스미", "20871711", 11500, 14000],
+    ["소친친", "36628482", 10500, 16000],
+    ["화화담", "1872323656", 7800, 17800],
+    ["고쿠텐", "1560991978", 10000, 16000],
+    ["환희", "1885645301", 17000, 28000],
+    ["혜화곱창", "1784269711", 18000, 38000],
+    ["봉심식당", "820059446", 16000, 22000],
+    ["칸다소바", "1228101335", 11000, 13500],
+    ["고기하다", "1773040934", 19000, 33000],
+    ["혜화돌쇠아저씨", "1290150733", 17900, 37500],
+    ["솥솥", "1005665891", 14000, 25000],
+    ["메밀향그집", "31751937", 26000, 52000],
+    ["버거파크", "13158227", 7400, 9400],
+    ["솔나무길된장예술", "11730630", 12000, 28000],
+    ["정돈", "37126807", 14000, 23000],
+    ["미도인", "1747236582", 10500, 17700],
+    ["다다미 소바", "12026377", 8500, 11000],
+    ["하이디라오", "1583520034", 20000, 50000],
+    ["피터팬스테이크", "34786204", 12000, 25000],
+    ["리본원도우", "1577816531", 10500, 14000],
+    ["삼겹살 전문점 대통령", "31403119", 17000, 58000],
+    ["순대실록", "20148711", 10000, 32000],
+    ["촌놈", "11867483", 16000, 75000],
+    ["오레노카츠", "1094922435", 11000, 18000],
+    ["프리모바치오바치", "35031115", 14000, 32000],
+    ["틴틴", "1083891202", 12900, 19900],
+    ["더 키친750", "37420979", 15900, 17900],
+    ["소바의온도", "1669386778", 9000, 14500],
+    ["에머이", "584785475", 10900, 18000],
+    ["서울프랑스", "1696688941", 15000, 23000],
+    ["고부기 멕시칸그릴", "1610379824", 9000, 42000],
+    ["우와", "1400319959", 16000, 18000]
+  ]
 }
